In String Gauge Calculator feature the recyclerview is implemented with adding/removing string. Updating the database and the business logic to calculate the string gauge is missing. Current UI update with databinding.

+++ app/src/main/java/com/example/alexiguitartuner/commons/domain/ChordTable.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class ChordTable(
-    @PrimaryKey(autoGenerate = true) val chordTableId : Int,
+    @PrimaryKey(autoGenerate = true) val chordTableId : Long,
     val chordId : Int,
     val position : Short,
     val pitchPos : List<Int>
